pyproject.toml
[tool.poetry]
name = "learning-security"
version = "0.1.0"
description = "Learning Fastapi JWT security"
authors = ["NwoluDavid <nwoludave@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-multipart = "^0.0.17"
pydantic = "^2.10.0"
pyjwt = "^2.10.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}

[tool.poetry.group.dev.dependencies]
fastapi = "^0.115.5"
uvicorn = "^0.32.1"
passlib = "^1.7.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
